+++ models/token.model.js
@@ -5,7 +5,7 @@ const sequelize = require('../system/core/db.connection');
 const Token = sequelize.define("Token",
     {
       id: {
-        type: DataTypes.INTEGER,
+        type: DataTypes.BIGINT(11),
         primaryKey: true,
         autoIncrement: true,
         allowNull: false
